+++ modules/Get-Provider/get-provider.js
@@ -1,0 +1,8 @@
+const tezsterManager = require("../../tezster-manager");
+
+async function getProvider() {
+    await tezsterManager.loadTezsterConfig(); 
+    console.log(tezsterManager.getProvider());
+}
+
+module.exports = { getProvider };
