+++ modules/List-Accounts/list-accounts.js
@@ -1,0 +1,16 @@
+const tezsterManager = require("../../tezster-manager");
+
+async function listAccounts() {  
+    await tezsterManager.loadTezsterConfig();
+    const config = tezsterManager.config;
+    if(Object.keys(config.accounts).length > 0){
+        for(var i in config.accounts){
+            console.log(tezsterManager.output(config.accounts[i].label + " - " + config.accounts[i].pkh + " (" + config.accounts[i].identity + ")"));
+        }
+    }
+    else{    
+        console.log(tezsterManager.outputError("No Account is available !!"));        
+    }
+}
+
+module.exports = { listAccounts };
