+++ modules/List-Contracts/list-contracts.js
@@ -1,0 +1,16 @@
+const tezsterManager = require("../../tezster-manager");
+
+async function listContracts() {
+    await tezsterManager.loadTezsterConfig();    
+    const config = tezsterManager.config;
+    if(Object.keys(config.contracts).length > 0){        
+        for(var i in config.contracts){
+            console.log(tezsterManager.output(config.contracts[i].label + " - " + config.contracts[i].pkh + " (" + config.contracts[i].identity + ")"));        
+        }
+    }
+    else{
+        console.log(tezsterManager.outputError("No Contracts are Available !!"));        
+    }
+}
+
+module.exports = { listContracts };
